Use inline dependency declaration in convert2onnx.py
